This is the first of several PRs cleaning up `tseries.offsets`.  The ultimate goals of this series of PRs are:

- Fix slow implementation of `DateOffset.__eq__`, as that gets called by `Period.__eq__`.
- Make `DateOffset` immutable, since it is attached to a `Period` object which is supposed to be immutable (TODO: fill in the appropriate GH issue)
- Move `tseries.offsets` into cython so that `_libs.period` and `_libs.tslib` can import it guilt-free and not need to do run-time imports.  See TODO comment in `_libs.__init__`:

```
# TODO
# period is directly dependent on tslib and imports python
# modules, so exposing Period as an alias is currently not possible
```

The biggest impediment to the immutability goal is the `kwds` attribute, which is just a `dict`.  The first couple of steps in this sequence is focused on whittling down the number of attributes set at runtime.

This PR is mainly fixing typos and removing redundant methods.


- [x] passes `git diff upstream/master -u -- "*.py" | flake8 --diff`