@pep8speaks 2017-08-23
Hello @jbrockmendel! Thanks for updating the PR.

 - In the file [`pandas/tseries/offsets.py`](https://github.com/pandas-dev/pandas/blob/3d1e6f8db2b2dd471ca0138b2eb65ca6ac1930e5/pandas/tseries/offsets.py), following are the PEP8 issues :

> [Line 593:35](https://github.com/pandas-dev/pandas/blob/3d1e6f8db2b2dd471ca0138b2eb65ca6ac1930e5/pandas/tseries/offsets.py#L593): [E261](https://duckduckgo.com/?q=pep8%20E261) at least two spaces before inline comment
> [Line 690:35](https://github.com/pandas-dev/pandas/blob/3d1e6f8db2b2dd471ca0138b2eb65ca6ac1930e5/pandas/tseries/offsets.py#L690): [E261](https://duckduckgo.com/?q=pep8%20E261) at least two spaces before inline comment
> [Line 960:35](https://github.com/pandas-dev/pandas/blob/3d1e6f8db2b2dd471ca0138b2eb65ca6ac1930e5/pandas/tseries/offsets.py#L960): [E261](https://duckduckgo.com/?q=pep8%20E261) at least two spaces before inline comment
> [Line 1451:35](https://github.com/pandas-dev/pandas/blob/3d1e6f8db2b2dd471ca0138b2eb65ca6ac1930e5/pandas/tseries/offsets.py#L1451): [E261](https://duckduckgo.com/?q=pep8%20E261) at least two spaces before inline comment
> [Line 1512:35](https://github.com/pandas-dev/pandas/blob/3d1e6f8db2b2dd471ca0138b2eb65ca6ac1930e5/pandas/tseries/offsets.py#L1512): [E261](https://duckduckgo.com/?q=pep8%20E261) at least two spaces before inline comment