@jbrockmendel 2017-09-14
I think this test error is unrelated.  Pls confirm.