@jbrockmendel 2017-09-22
> can you run the freq asv's to make sure nothing has changed here.

Begrudgingly...
```
$ asv continuous -f 1.1 -E virtualenv master HEAD -b freq
[...]
       before           after         ratio
     [2bec750b]       [d1e9161b]
+          72.9ms            126ms     1.73  timeseries.DatetimeIndex.time_infer_freq_none
-          57.9ms           45.0ms     0.78  timeseries.DatetimeIndex.time_infer_freq_business
-          44.7ms           34.3ms     0.77  timeseries.DatetimeIndex.time_infer_freq_daily

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.

$ asv continuous -f 1.1 -E virtualenv master HEAD -b freq
[...]
     [2bec750b]       [d1e9161b]
+          31.7ms           43.8ms     1.38  timeseries.DatetimeIndex.time_infer_freq_business
+     1.52±0.04μs      1.99±0.05μs     1.30  timestamp.TimestampProperties.time_freqstr
+          43.1ms           50.4ms     1.17  timeseries.DatetimeIndex.time_infer_freq_none

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.

$ asv continuous -f 1.1 -E virtualenv master HEAD -b freq
[...]
     [2bec750b]       [d1e9161b]
+          25.6ms           30.7ms     1.20  timeseries.DatetimeIndex.time_infer_freq_business
-          27.2ms           24.3ms     0.89  timeseries.DatetimeIndex.time_infer_freq_daily

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.

$ asv continuous -f 1.1 -E virtualenv master HEAD -b freq
[...]
      before           after         ratio
     [2bec750b]       [d1e9161b]
-          60.6ms           53.9ms     0.89  timeseries.DatetimeIndex.time_infer_freq_none
-          34.6ms           28.7ms     0.83  timeseries.DatetimeIndex.time_infer_freq_business
-          30.9ms           23.4ms     0.76  timeseries.DatetimeIndex.time_infer_freq_daily
-      2.87±0.3μs      2.14±0.07μs     0.75  timestamp.TimestampProperties.time_freqstr

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.
```