@jreback 2017-09-22
This is repeatable.
```
    before     after       ratio
  [f797c1dc] [d1e9161b]
+  529.00μs    12.36ms     23.36  timeseries.DatetimeIndex.time_infer_freq_business
+   20.68ms    24.21ms      1.17  timeseries.DatetimeIndex.time_infer_freq_none

SOME BENCHMARKS HAVE CHANGED SIGNIFICANTLY.

```