@jbrockmendel 2017-09-23
After updating, I get the same timings too.  But it looks like this may be due to a bug in master.  Manually stepping through the benchmark, `timeseries.DatetimeIndex` includes the code

```
rng8 = date_range(start='1/1/1700', freq='B', periods=100000)
```
which for me (py27, ubuntu...) is giving:

```
OverflowError: Python int too large to convert to C long
Exception OverflowError: 'Python int too large to convert to C long' in 'pandas._libs.tslib._delta_to_nanoseconds' ignored
```
and the resulting `rng8` object is a `DatetimeIndex` object with 5 entries.  Before anything else, can you reproduce this?