@jbrockmendel 2017-09-23
git bisect tells me the first bad commit was b59f107a529ac9b1e2859217ea58f7e26b6749d0