@jreback 2017-09-23
let me see if I can work around: https://github.com/pandas-dev/pandas/issues/17637