@jbrockmendel 2017-09-23
After #17637, benchmarks are now unaffected.