@jreback 2017-09-23
> After #17637, benchmarks are now unaffected.

does this mean they are the same (worse) or the same as master?