@jbrockmendel 2017-09-23
```
asv continuous -f 1.1 -E virtualenv master HEAD -b freq
[...]
BENCHMARKS NOT SIGNIFICANTLY CHANGED.
```