@jbrockmendel 2017-09-23
Just pushed.  Besides the requested docstring, had to edit the asv to avoid the new OverflowError.