@jreback 2017-09-23
lgtm. ping on green.

side note, I believe you can make ``_params`` readonly_cached if i read it correctly (another PR of course :>)