@jbrockmendel 2017-09-23
> side note, I believe you can make _params readonly_cached if i read it correctly (another PR of course :>)

That is pretty much the original motivation here.  There are a couple more steps between here and there, since until we get `kwds` locked down (see WIP #17458) and make relevant attrs immutable, cache invalidation is a PITA.